+++ src/love/event.h
@@ -1,6 +1,8 @@
 #ifndef SRC_LOVE_EVENT_H_
 #define SRC_LOVE_EVENT_H_
 
+#include <SDL.h>
+
 namespace love {
 /**
  * Manages events, like keypresses.
@@ -11,7 +13,13 @@ class event {
 	 * Exits the ChaiLove application.
 	 */
 	void quit();
+	void update();
+
 	bool m_shouldclose = false;
+
+
+	private:
+	SDL_Event sdlEvent;
 };
 
 }  // namespace love